Put all the stuff in index
